+++ src/rest/types/ignite.ts
@@ -160,6 +160,14 @@ export interface Container {
 	};
 
 	/**
+	 * Overrides that were provided manually to the container
+	 */
+
+	overrides: {
+		resources?: Partial<Resources>;
+	} | null;
+
+	/**
 	 * The type of this container
 	 */
 	type: RuntimeType;
@@ -247,6 +255,23 @@ export interface Deployment {
 	 * Metadata for deployment
 	 */
 	metadata: DeploymentMetaData | null;
+
+	/**
+	 * Build cache settings for deployment
+	 */
+	build_cache_enabled: boolean;
+
+	/**
+	 * Build settings for deployment
+	 */
+	build_settings?: BuildSettings;
+}
+
+export interface BuildSettings {
+	/**
+	 * Root directory for build
+	 */
+	root_directory?: string;
 }
 
 export interface DeploymentMetaData {
@@ -787,7 +812,11 @@ export type IgniteEndpoints =
 			CreateDeploymentConfig
 	  >
 	| Endpoint<'DELETE', '/v1/ignite/deployments/:deployment_id', Empty>
-	| Endpoint<'DELETE', '/v1/ignite/containers/:container_id', Empty>
+	| Endpoint<
+			'DELETE',
+			'/v1/ignite/containers/:container_id',
+			Empty | {container: Container}
+	  >
 	| Endpoint<
 			'GET',
 			'/v1/ignite/containers/:container_id/logs',