+++ CHANGELOG.md
@@ -1,5 +1,29 @@
 # @onehop/js
 
+## 1.20.1
+
+### Patch Changes
+
+- 2a64d7b: Bump
+
+## 1.20.0
+
+### Minor Changes
+
+- 527b63b: Fix build settings type
+
+## 1.19.0
+
+### Minor Changes
+
+- 1e0145e: Add build_settings and build_cache_enabled to deployment
+
+## 1.18.0
+
+### Minor Changes
+
+- 93fe56b: Add overrides to container object
+
 ## 1.17.0
 
 ### Minor Changes